fix: ref 句柄改 64 位、JsRuntime 用 Mutex 互斥并安全关闭
- ref 低 32 位 slot 下标、高 32 位 generation，回绕在实践中不可能；isValid 联动引擎关闭状态
- JsRuntime 互斥来自内部 Mutex，不再依赖单车道 dispatcher；shutdown/close 先中断再在锁下关闭，
  引擎被占用时 close 延后到当前工作结束
- README 示例补齐 engine 作用域与超时处理；新增多车道 dispatcher、shutdown、busy close 回归测试

+++ mquickjs-core/src/androidMain/kotlin/wang/harlon/mquickjs/NativeBridge.kt
@@ -15,13 +15,13 @@ internal object NativeBridge {
     @JvmStatic external fun nativeEval(ptr: Long, code: ByteArray, fileName: ByteArray, flags: Int): NativeValue?
     @JvmStatic external fun nativeDefineFunction(ptr: Long, name: ByteArray, id: Int, flags: Int): NativeValue?
     @JvmStatic external fun nativeInterrupt(ptr: Long)
-    @JvmStatic external fun nativeRefRetain(ptr: Long, ref: Int)
-    @JvmStatic external fun nativeRefRelease(ptr: Long, ref: Int)
-    @JvmStatic external fun nativeRefGet(ptr: Long, ref: Int, name: ByteArray, flags: Int): NativeValue?
-    @JvmStatic external fun nativeRefGetIndex(ptr: Long, ref: Int, index: Int, flags: Int): NativeValue?
-    @JvmStatic external fun nativeRefSet(ptr: Long, ref: Int, name: ByteArray, value: NativeValue): NativeValue?
-    @JvmStatic external fun nativeRefCall(ptr: Long, ref: Int, thisRef: Int, args: Array<NativeValue>, flags: Int): NativeValue?
-    @JvmStatic external fun nativeRefToJson(ptr: Long, ref: Int): NativeValue?
+    @JvmStatic external fun nativeRefRetain(ptr: Long, ref: Long)
+    @JvmStatic external fun nativeRefRelease(ptr: Long, ref: Long)
+    @JvmStatic external fun nativeRefGet(ptr: Long, ref: Long, name: ByteArray, flags: Int): NativeValue?
+    @JvmStatic external fun nativeRefGetIndex(ptr: Long, ref: Long, index: Int, flags: Int): NativeValue?
+    @JvmStatic external fun nativeRefSet(ptr: Long, ref: Long, name: ByteArray, value: NativeValue): NativeValue?
+    @JvmStatic external fun nativeRefCall(ptr: Long, ref: Long, thisRef: Long, args: Array<NativeValue>, flags: Int): NativeValue?
+    @JvmStatic external fun nativeRefToJson(ptr: Long, ref: Long): NativeValue?
 
     @JvmStatic
     fun onHostCall(target: Any, id: Int, args: Array<NativeValue?>): NativeValue {
